Improve separation of streaming and batch APIs. (#57)
* separated detection objects used in both batch and streaming into their own header file, and removed streaming detection component type from the interface header file.
